requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
SQLAlchemy==2.0.43
typing_extensions==4.14.1
Werkzeug==3.1.3
